Windows_en
Metasploit post module for Windows enumeration.As a lazy person I always want to make things easy and automatic  I created this module just for saving time. This module will run the bunch of WMIC commands and save the output in a text file and it will also run the others Post modules.